added functions

+++ lists.h
@@ -5,15 +5,15 @@
 #include <stdlib.h>
 
 template <class T>
-struct Node {
+struct LinkedListNode {
     public:
         T value;
-        struct Node* next;
+        struct LinkedListNode* next;
 
-        Node(T value) {
+        LinkedListNode(T value) {
             this->value = value;
         }
-        ~Node() {
+        ~LinkedListNode() {
             if (this->next != NULL)
                 free(this->next);
         }
@@ -23,13 +23,14 @@ template <class T>
 struct LinkedList {
     private:
         int length = 0;
-        struct Node<T> *start   = NULL,
+        struct LinkedListNode<T> *start   = NULL,
                        *end     = NULL;
 
     public:
         LinkedList() {
         }
         ~LinkedList() {
+            this->clear();
         }
 
         int size() {
@@ -39,27 +40,32 @@ struct LinkedList {
             if (index >= length) {
                 this->append(v);
             } else {
-                struct Node<T>* temp = this->start;
+                struct LinkedListNode<T> *n_start = (struct LinkedListNode<T>*) malloc(sizeof(struct LinkedListNode<T>)),
+                                         *temp = n_start;
+                n_start->next = this->start;
+
                 for (int i = 0; i < index; i++)
                     temp = temp->next;
 
-                struct Node<T>* t = (struct Node<T>*) malloc(sizeof(struct Node<T>));
+                struct LinkedListNode<T>* t = (struct LinkedListNode<T>*) malloc(sizeof(struct LinkedListNode<T>));
                 t->value = v;
                 t->next = temp->next;
                 temp->next = t;
+
+                this->start = n_start->next;
             }
 
             this->length++;
         }
         void append(T v) {
             if (length == 0) {
-                this->start = (struct Node<T>*) malloc(sizeof(struct Node<T>));
+                this->start = (struct LinkedListNode<T>*) malloc(sizeof(struct LinkedListNode<T>));
                 this->start->value = v;
                 this->start->next = NULL;
 
                 this->end = this->start;
             } else {
-                this->end->next = (struct Node<T>*) malloc(sizeof(struct Node<T>));
+                this->end->next = (struct LinkedListNode<T>*) malloc(sizeof(struct LinkedListNode<T>));
                 this->end->next->value = v;
                 this->end->next->next = NULL;
 
@@ -72,7 +78,7 @@ struct LinkedList {
             if (index > this->length || this->length == 0)
                 return NULL;
 
-            struct Node<T>* target = this->start;
+            struct LinkedListNode<T>* target = this->start;
             for (int i = 0; i < index; i++) {
                 target = target->next;
             }
@@ -83,7 +89,7 @@ struct LinkedList {
             if (index > length || length == 0)
                 return;
 
-            struct Node<T>* target = this->start;
+            struct LinkedListNode<T>* target = this->start;
             for (int i = 0; i < index-1; i++) {
                 target = target->next;
             }
@@ -91,10 +97,27 @@ struct LinkedList {
             target->next = target->next->next;
             this->length--;
         }
-        void clear() {
-            deleteAll(this->start);
+        void removeValue(T v) {
+            struct LinkedListNode<T> *n_start = (struct LinkedListNode<T>*) malloc(sizeof(struct LinkedListNode<T>)),
+                                     *cu = n_start;
+            n_start->next = this->start;
+
+            while (cu->next != NULL) {
+                if (cu->next->value == v) {
+                    cu->next = cu->next->next;
+                    this->length--;
+                } else {
+                    cu = cu->next;
+                }
+            }
+
+            this->start = n_start->next;
         }
-        void deleteAll(struct Node<T>* cu) {
+        void clear() {
+            if (this->length>0)
+                deleteAll(this->start);
+        }
+        void deleteAll(struct LinkedListNode<T>* cu) {
             if (cu->next != NULL) {
                 deleteAll(cu->next);
             }
@@ -102,7 +125,7 @@ struct LinkedList {
             this->length--;
         }
         int index(T v) {
-            struct Node<T>* cu = this->start;
+            struct LinkedListNode<T>* cu = this->start;
             int i = 0;
 
             while (cu != NULL && cu->value != v) {
@@ -115,7 +138,7 @@ struct LinkedList {
             return i;
         }
         bool contains(T v) {
-            struct Node<T>* cu = this->start;
+            struct LinkedListNode<T>* cu = this->start;
             while(cu->next != NULL && cu->value != v) {
                 cu = cu->next;
             }
@@ -125,10 +148,10 @@ struct LinkedList {
             return true;
         }
         void sort(bool (*comp)(T, T)) {
-            struct Node<T>* temp_n = (struct Node<T>*) malloc(sizeof(struct Node<T>));
+            struct LinkedListNode<T>* temp_n = (struct LinkedListNode<T>*) malloc(sizeof(struct LinkedListNode<T>));
             temp_n->next = this->start;
 
-            struct Node<T> *co = temp_n,
+            struct LinkedListNode<T> *co = temp_n,
                            *be_cu = this->start;
 
             for (int i = 1; i < this->length; i++) {
@@ -137,10 +160,10 @@ struct LinkedList {
 
                 for (int j = 0; j < i; j++) {
                     if (!comp(co->next->value, be_cu->next->value)) {
-                        struct Node<T>* h = be_cu->next;
+                        struct LinkedListNode<T>* h = be_cu->next;
                         be_cu->next = be_cu->next->next;
 
-                        struct Node<T>* h2 = co->next;
+                        struct LinkedListNode<T>* h2 = co->next;
                         co->next = h;
                         h->next = h2;
 
@@ -156,11 +179,46 @@ struct LinkedList {
 
             this->start = temp_n->next;
         }
+        void filter(bool (*fi)(T)) {
+            struct LinkedListNode<T> *n_start = (struct LinkedListNode<T>*) malloc(sizeof(struct LinkedListNode<T>)),
+                           *cu = n_start;
+            cu->next = this->start;
 
+            while (cu->next != NULL) {
+                if (!fi(cu->next->value)) {
+                    cu->next = cu->next->next;
+                    this->length--;
+                } else {
+                    cu = cu->next;
+                }
+            }
+
+            this->start = n_start->next;
+        }
+        void map(T (*m)(T)) {
+            struct LinkedListNode<T>* cu = this->start;
+            while (cu != NULL) {
+                cu->value = m(cu->value);
+                cu = cu->next;
+            }
+        }
+        T reduce(T (*r)(T, T)) {
+            struct LinkedListNode<T>* cu = this->start->next;
+            if (this->length == 0)
+                return NULL;
+
+            T ret = this->start->value;
+            while (cu != NULL) {
+                ret = r(ret, cu->value);
+                cu = cu->next;
+            }
+
+            return ret;
+        }
         std::string toString() {
             std::stringstream ret;
             ret << "[";
-            struct Node<T>* target = this->start;
+            struct LinkedListNode<T>* target = this->start;
             for (int i = 0; i < this->length; i++) {
                 ret << target->value << (i < this->length-1 ? ", " : "");
                 target = target->next;
